Mover "hover from reference" code to a separate file
